+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/assets/AssetAsPlannedControllerAllIT.java
@@ -85,12 +85,11 @@ class AssetAsPlannedControllerAllIT extends IntegrationTestSpecification {
         assetsSupport.defaultAssetsAsPlannedStored();
 
         //THEN filter=owner,EQUAL,OWN
-        final String filterOperator = "AND";
+        final String filter = "owner,EQUAL," + ownerValue + ",AND";
         given()
                 .header(oAuth2Support.jwtAuthorization(ADMIN))
                 .contentType(ContentType.JSON)
-                .queryParam("filter", "owner,EQUAL," + ownerValue)
-                .queryParam("filterOperator", filterOperator)
+                .queryParam("filter", filter)
                 .when()
                 .get("/api/assets/as-planned")
                 .then()