+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/qualitynotification/alert/AlertControllerFilterIT.java
@@ -1,0 +1,104 @@
+/********************************************************************************
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+package org.eclipse.tractusx.traceability.integration.qualitynotification.alert;
+
+import io.restassured.http.ContentType;
+import org.eclipse.tractusx.traceability.integration.IntegrationTestSpecification;
+import org.eclipse.tractusx.traceability.integration.common.support.AlertNotificationsSupport;
+import org.hamcrest.Matchers;
+import org.jose4j.lang.JoseException;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import java.time.LocalDate;
+
+import static io.restassured.RestAssured.given;
+import static org.eclipse.tractusx.traceability.common.security.JwtRole.ADMIN;
+
+
+class AlertControllerFilterIT extends IntegrationTestSpecification {
+    @Autowired
+    AlertNotificationsSupport alertNotificationsSupport;
+
+    @Test
+    void givenAlerts_whenProvideNoFilter_thenReturnAll() throws JoseException {
+        // given
+        alertNotificationsSupport.defaultAlertsStored();
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .get("/api/alerts")
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("totalItems", Matchers.is(14))
+                .body("content", Matchers.hasSize(10));
+    }
+
+    @Test
+    void givenAlerts_whenProvideBpnFilter_thenReturnExpectedResult() throws JoseException {
+        // given
+        alertNotificationsSupport.defaultAlertsStored();
+        String filter = "?filter=bpn,STARTS_WITH,BPNL00000001OWN,OR";
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .get("/api/alerts" + filter)
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("content", Matchers.hasSize(8))
+                .body("totalItems", Matchers.is(8));
+    }
+
+    @Test
+    void givenAlerts_whenProvideBpnFilterAnd_thenReturnExpectedResult() throws JoseException {
+        // given
+        alertNotificationsSupport.defaultAlertsStored();
+        String filter = "?filter=bpn,STARTS_WITH,BPNL00000001OWN,AND&filter=createdDate,AT_LOCAL_DATE,2023-10-10,AND";
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .get("/api/alerts" + filter)
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("content", Matchers.hasSize(2))
+                .body("totalItems", Matchers.is(2));
+    }
+}